requirements.txt
absl-py==0.5.0
astor==0.7.1
bleach==1.5.0
cloudpickle==0.5.6
cycler==0.10.0
dask==0.19.2
decorator==4.3.0
gast==0.2.0
grpcio==1.15.0
h5py==2.8.0
html5lib==0.9999999
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
Markdown==3.0.1
matplotlib==3.0.0
networkx==2.2
numpy==1.15.2
opencv-python==3.4.3.18
Pillow==5.2.0
protobuf==3.6.1
pyparsing==2.2.1
python-dateutil==2.7.3
PyWavelets==1.0.1
scikit-image==0.14.0
scikit-learn==0.20.0
scipy==1.1.0
six==1.11.0
tensorboard==1.8.0
tensorflow-gpu==1.8.0
termcolor==1.1.0
toolz==0.9.0
Werkzeug==0.14.1
